+++ CHANGELOG.md
@@ -15,6 +15,7 @@ accidentally triggering the load of a previous DB version.**
 * #5396 Fix SEGMENTBY columns predicates to be pushed down
 * #5410 Fix file trailer handling in the COPY fetcher
 * #5233 Out of on_proc_exit slots on guc license change
+* #5427 Handle user-defined FDW options properly
 
 **Thanks**
 * @nikolaps for reporting an issue with the COPY fetcher